refactor: related socks utils
